Corrected varsion and added changelog
