+++ androidExample/src/main/java/dev/snipme/androidexample/MainActivity.kt
@@ -32,8 +32,8 @@ import dev.snipme.highlights.model.SyntaxLanguage
 import dev.snipme.highlights.model.SyntaxTheme
 import dev.snipme.highlights.model.SyntaxThemes
 import dev.snipme.highlights.model.SyntaxThemes.useDark
-import dev.snipme.kodeview.view.CodeEditText
 import dev.snipme.kodeview.view.CodeTextView
+import dev.snipme.kodeview.view.material3.CodeEditText
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
