+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ kotlinAndroid = "1.9.21"
 kotlinMultiplatform = "1.9.21"
 compose = "1.5.11"
 androidLibrary = "8.1.1"
-kodeview = "0.7.0"
+kodeview = "0.8.0"
 highlights = "0.7.1"
 composeMaterial = "1.2.1"
